fix: Allow training on a function with 1D domain
Training fails for a function on scalars.
This is because in this case numpy outputs a scalar covariance
instead of a 1x1 matrix.  This in turn breaks the calculation of the
Cholesky decomposition matrix.

Fix this by checking the dimensions of the training data and handling
this case specially.

Should check more formally that a Cholesky decomposition and covarianc
matrix even make sense in this case.