refactor: Split out the Cholesky emulator
Split the Cholesky nearest neighbour emulator into a separate class.
This keeps the logic concerned with emulation and learning as you go
separate.

In future there will be multiple options for the emulator used, while
the learn-as-you-go logic remains constant.

Drop support for Python 3.5 by including type annotation for the
callables defined by the emulator.