fix: No duplicate values in initial training set
The initial values of `leaner.frac_err_local` and
`learner.abs_err_local` were set to 0.0.  This introduces a subtle bug:
the evaluator believes that values from the true function have an
unacceptable err and re-evaluates the true function.  This has the
unfortunate side effect of adding the corresponding x and y values to
the batch training sets again.  This in turn results in the emulator
initially being trained after half the steps expected.

Set `leaner.frac_err_local = 1.0` and `learner.abs_err_local = 0.05`.
These are the default values later re-set by the training method.

Track the number of exact emulations more carefully.

Add more checks to the tests of the initial training.