feat: Retrain when 50% more points are known
Set the retraining threshold dynamically.  In particular, retrain when
frac_cv more points are in the training set than when the emulator was
last trained.

Remove the `Learner.other_train_thresh`.  This is only useful for fixed
retraining thresholds.

Remove `Learner.overrideDefaults`.  This was used for setting different
training thresholds, but they can be set by modifying attributes
directly.