feat: Add API to allow updating emulator's data
Add a method `BaseEmulator.add_data` to add data to an emulator.  Use
this in `Learner` to provide data to the emulator on each call to
`Learner.eval_true_func` and to put cross validation data back in after
training the error model.

The method is not yet implemented in a useful way in any emulators.